python/100_days_of_code/pyproject.toml
[tool.poetry]
name = "100-days-of-code"
version = "0.1.0"
description = ""
authors = ["Saeed Gholami <saeed.ghollami@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
pillow = "^10.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
